+++ langs/calc/CalcParser.scala
@@ -18,7 +18,7 @@ import cats.syntax.all.given
 
 import forja.*
 import forja.dsl.*
-import forja.src.Reader
+import forja.source.Reader
 import forja.wf.Wellformed
 
 import lang.*
