+++ langs/calc/CalcReader.scala
@@ -19,7 +19,7 @@ import cats.syntax.all.given
 import forja.*
 import forja.Builtin.{Error, SourceMarker}
 import forja.dsl.*
-import forja.src.{Reader, SourceRange}
+import forja.source.{Reader, SourceRange}
 import forja.wf.Wellformed
 
 import lang.*
