+++ src/sexpr/SExprReader.test.scala
@@ -17,7 +17,7 @@ package forja.sexpr
 import forja.*
 import forja.Builtin.{Error, SourceMarker}
 import forja.sexpr.lang.{SAtom, SList}
-import forja.src.{Source, SourceRange}
+import forja.source.{Source, SourceRange}
 
 class SExprReaderTests extends munit.FunSuite:
   extension (str: String)
